pyproject.toml
[project]
name = "sjg-lms-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.6.0",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "nicegui>=2.24.0",
    "pandas>=2.3.2",
    "pdoc3>=0.11.6",
    "python-docx>=1.2.0",
    "qrcode>=8.2",
]
